V.0.0.2.001 : Fix.
- Removed all byte convertion.
- Some value have been replaced by enum.
- MapData no longuer care of tile specifique properties, it add or remove tile only.
- MapData no longuer have his own appState, HexGridManager got the main reference.
- All listener are encapsulated, o longuer class implementation. (W.I.P)
- MouseEvent now use only one method for input, event contain the mouse input used by the user.

+++ HexGrid_Editor/src/gui/FileManagerPopup.java
@@ -56,7 +56,7 @@ public class FileManagerPopup extends DialogWindow {
             }
         };
         File[] flist = folder.listFiles();
-        byte i = 0;
+        int i = 0;
         for (int j = 0; j < flist.length; j++) {
             if (flist[j].getName().equalsIgnoreCase("Reset") || flist[j].getName().equalsIgnoreCase("Temp")) {
                 i++;
